+++ resources/forms-ui/src/app/app-routing.module.ts
@@ -106,6 +106,11 @@ import { ExecViewSuperExectivesPageComponent } from './pages/dashboard/executive
 import { ExecViewBranchExectivesPageComponent } from './pages/dashboard/executive/exec-view-branch-exectives-page/exec-view-branch-exectives-page.component';
 import { ClientFavoriteFormsPageComponent } from './pages/dashboard/client/client-favorite-forms-page/client-favorite-forms-page.component';
 import { ClientSuggestMerchantPageComponent } from './pages/dashboard/client/client-suggest-merchant-page/client-suggest-merchant-page.component';
+import { ClientDeletedFormsPageComponent } from './pages/dashboard/client/client-deleted-forms-page/client-deleted-forms-page.component';
+import { ExecViewAccountDetailsPageComponent } from './pages/dashboard/executive/exec-view-account-details-page/exec-view-account-details-page.component';
+import { CreateSectorPageComponent } from './pages/dashboard/git-admin/create-sector-page/create-sector-page.component';
+import { EditSectorPageComponent } from './pages/dashboard/git-admin/edit-sector-page/edit-sector-page.component';
+import { ViewSectorListPageComponent } from './pages/dashboard/git-admin/view-sector-list-page/view-sector-list-page.component';
 
 const routes: Routes = [
   {
@@ -197,6 +202,10 @@ const routes: Routes = [
           {
             path: 'section',
             component: CreateSectionPageComponent
+          },
+          {
+            path: 'sector',
+            component: CreateSectorPageComponent
           }
         ]
       },
@@ -223,6 +232,10 @@ const routes: Routes = [
           {
             path: 'section',
             component: EditSectionPageComponent
+          },
+          {
+            path: 'sector',
+            component: EditSectorPageComponent
           },
           {
             path: 'access_code',
@@ -265,6 +278,10 @@ const routes: Routes = [
           {
             path: 'branch',
             component: ViewBranchListsPageComponent
+          },
+          {
+            path: 'sector',
+            component: ViewSectorListPageComponent
           },
           {
             path: 'front_desk',
@@ -351,6 +368,11 @@ const routes: Routes = [
         path: 'favorites',
         canActivate: [AuthGuard],
         component: ClientFavoriteFormsPageComponent
+      },
+      {
+        path: 'deleted',
+        canActivate: [AuthGuard],
+        component: ClientDeletedFormsPageComponent
       },
       {
         path: 'printing',
@@ -467,6 +489,10 @@ const routes: Routes = [
       {
         path: 'accounts',
         component: ExecAccountsListPageComponent
+      },
+      {
+        path: 'account_details',
+        component: ExecViewAccountDetailsPageComponent
       },
       {
         path: 'branches',