+++ resources/forms-ui/src/app/components/navigation-bar/navigation-bar.component.ts
@@ -1,8 +1,8 @@
-import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { Component, OnInit } from '@angular/core';
+import { UserTypes } from 'src/app/enums/user-types.enum';
 import { AccountService } from 'src/app/services/account/account.service';
 import { LocalStorageService } from 'src/app/services/storage/local-storage.service';
-import { UserTypes } from 'src/app/enums/user-types.enum';
 
 @Component({
   selector: 'app-navigation-bar',