+++ resources/forms-ui/src/app/pages/dashboard/client/client-forms-history-page/client-forms-history-page.component.ts
@@ -7,7 +7,6 @@ import { ClientService } from 'src/app/services/client/client.service';
 import { EndpointService } from 'src/app/services/endpoint/endpoint.service';
 import { DateTimeService } from 'src/app/services/date-time/date-time.service';
 import { ReloadingService } from 'src/app/services/reloader/reloading.service';
-import { DownloaderService } from 'src/app/services/downloader/downloader.service';
 import { Component, OnInit, ViewChild, TemplateRef, OnDestroy } from '@angular/core';
 import { LocalStorageService } from 'src/app/services/storage/local-storage.service';
 
@@ -20,6 +19,7 @@ export class ClientFormsHistoryPageComponent implements OnInit, OnDestroy {
   form: any;
   user: Users;
   query: string;
+  status: number;
   endDate: string;
   hasMore: boolean;
   hasData: boolean;
@@ -49,9 +49,9 @@ export class ClientFormsHistoryPageComponent implements OnInit, OnDestroy {
     private clientService: ClientService,
     private dateService: DateTimeService,
     private endpointService: EndpointService,
-    private downloadService: DownloaderService,
     private localStorageService: LocalStorageService
   ) {
+    this.status = 0;
     this.searchOption = '';
     this.rejectionNote = '';
     this.historyCollection = [];
@@ -100,7 +100,10 @@ export class ClientFormsHistoryPageComponent implements OnInit, OnDestroy {
   }
 
   showAll() {
+    this.status = 0;
     this.filterState = 'all';
+    this.historyCollection = [];
+    this.allHistoryCollection = [];
     this.getAllHistory();
   }
 
@@ -130,10 +133,11 @@ export class ClientFormsHistoryPageComponent implements OnInit, OnDestroy {
   }
 
   showProcessed() {
-    this.filterState = 'processed';
+    this.status = 2;
     this.loading = true;
     this.historyCollection = [];
     this.allHistoryCollection = [];
+    this.filterState = 'processed';
     this.clientService.getFormByStatus(_.toString(this.user.id), 2).then(
       forms => {
         this.hasMore = this.checkIfHasMore();
@@ -161,10 +165,11 @@ export class ClientFormsHistoryPageComponent implements OnInit, OnDestroy {
   }
 
   showProcessing() {
-    this.filterState = 'processing';
+    this.status = 1;
     this.loading = true;
     this.historyCollection = [];
     this.allHistoryCollection = [];
+    this.filterState = 'processing';
     this.clientService.getFormByStatus(_.toString(this.user.id), 1).then(
       forms => {
         this.hasMore = this.checkIfHasMore();
@@ -193,8 +198,9 @@ export class ClientFormsHistoryPageComponent implements OnInit, OnDestroy {
   }
 
   showDrafts() {
-    this.filterState = 'drafts';
+    this.status = 4;
     this.loading = true;
+    this.filterState = 'drafts';
     this.historyCollection = [];
     this.allHistoryCollection = [];
     this.clientService.getFormByStatus(_.toString(this.user.id), 4).then(
@@ -226,9 +232,10 @@ export class ClientFormsHistoryPageComponent implements OnInit, OnDestroy {
   }
 
   showSubmitted() {
-    this.filterState = 'submitted';
+    this.status = 0;
     this.loading = true;
     this.historyCollection = [];
+    this.filterState = 'submitted';
     this.allHistoryCollection = [];
     this.clientService.getFormByStatus(_.toString(this.user.id), 0).then(
       forms => {
@@ -259,8 +266,9 @@ export class ClientFormsHistoryPageComponent implements OnInit, OnDestroy {
   }
 
   showRejected() {
-    this.filterState = 'rejected';
+    this.status = 3;
     this.loading = true;
+    this.filterState = 'rejected';
     this.historyCollection = [];
     this.allHistoryCollection = [];
     this.clientService.getFormByStatus(_.toString(this.user.id), 3).then(
@@ -291,8 +299,9 @@ export class ClientFormsHistoryPageComponent implements OnInit, OnDestroy {
   }
 
   showReversed() {
-    this.filterState = 'reversed';
+    this.status = 5;
     this.loading = true;
+    this.filterState = 'reversed';
     this.historyCollection = [];
     this.allHistoryCollection = [];
     this.clientService.getFormByStatus(_.toString(this.user.id), 5).then(
@@ -346,7 +355,7 @@ export class ClientFormsHistoryPageComponent implements OnInit, OnDestroy {
   searchByFormCode() {
     this.loading = true;
     console.log('form code searching is running...');
-    this.clientService.findFormsInHistoryByCode(this.user.id.toString(), this.query).then(
+    this.clientService.findFormsInHistoryByCode(this.user.id.toString(), this.query, this.status).then(
       forms => {
         if (forms.length == 0) {
           this.loading = false;
@@ -356,6 +365,7 @@ export class ClientFormsHistoryPageComponent implements OnInit, OnDestroy {
           this.loading = false;
           this.foundNoForm = false;
           _.forEach(forms, (form) => {
+            form.logo = this.endpointService.storageHost + form.logo;
             this.historyCollection.push(form);
           });
         }
@@ -370,7 +380,7 @@ export class ClientFormsHistoryPageComponent implements OnInit, OnDestroy {
   searchByFormName() {
     console.log('form name searching is running ...');
     this.loading = true;
-    this.clientService.findFormsInHistoryByName(this.user.id.toString(), this.query).then(
+    this.clientService.findFormsInHistoryByName(this.user.id.toString(), this.query, this.status).then(
       forms => {
         if (forms.length == 0) {
           this.loading = false;
@@ -380,6 +390,7 @@ export class ClientFormsHistoryPageComponent implements OnInit, OnDestroy {
           this.loading = false;
           this.foundNoForm = false;
           _.forEach(forms, (form) => {
+            form.logo = this.endpointService.storageHost + form.logo;
             this.historyCollection.push(form);
           });
         }
@@ -404,6 +415,7 @@ export class ClientFormsHistoryPageComponent implements OnInit, OnDestroy {
           this.loading = false;
           this.foundNoForm = false;
           _.forEach(forms, (form) => {
+            form.logo = this.endpointService.storageHost + form.logo;
             this.historyCollection.push(form);
           });
         }
@@ -435,6 +447,7 @@ export class ClientFormsHistoryPageComponent implements OnInit, OnDestroy {
           this.loading = false;
           this.foundNoForm = false;
           _.forEach(forms, (form) => {
+            form.logo = this.endpointService.storageHost + form.logo;
             this.historyCollection.push(form);
           });
         }