+++ resources/forms-ui/src/app/pages/dashboard/git-admin/create-sector-page/create-sector-page.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { CreateSectorPageComponent } from './create-sector-page.component';
+
+describe('CreateSectorPageComponent', () => {
+  let component: CreateSectorPageComponent;
+  let fixture: ComponentFixture<CreateSectorPageComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ CreateSectorPageComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(CreateSectorPageComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});