+++ resources/forms-ui/src/app/pages/dashboard/git-admin/view-access-code-page/view-access-code-page.component.ts
@@ -1,7 +1,7 @@
-import { Component, OnInit, ViewChild, TemplateRef } from '@angular/core';
 import * as _ from 'lodash';
 import { Router } from '@angular/router';
 import { NgbModalRef, NgbModal } from '@ng-bootstrap/ng-bootstrap';
+import { Component, OnInit, ViewChild, TemplateRef } from '@angular/core';
 import { AccountService } from 'src/app/services/account/account.service';
 import { DateTimeService } from 'src/app/services/date-time/date-time.service';